Remove duplicate code and move some util functions from classes.py and rename it to relationship_manager.py
